pyproject.toml
[tool.poetry]
name = "b-12"
version = "0.1.0"
description = "for b-12 dev."
authors = ["brighto <jychoi1996@gmail.com>"]
readme = "README.md"
packages = [{include = "b_12"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
django-nextjs = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
